Fix bug in write_values for rectangular grids
The `write_values()` function read the `obstacles` array incorrectly in this version of LBM. For square grids, as the obstacles input was symmetric is did not matter. For non-square (rectangular) grids where `nx < ny`, the stride did not go out of bounds, however did not check the report entry in the array. For `nx > ny` this risked an out of bounds memory access (segmentation fault).